fix(dashboard): moving common charts to components folder
