+++ src/components/charts/area-chart.scss
