+++ src/components/charts/bar-chart.scss
