+++ src/pages/dashboard/_StateCumulativeDeathsContainer.js
@@ -1,7 +1,7 @@
 import React, { useMemo, useState } from 'react'
 import { graphql, useStaticQuery } from 'gatsby'
-import HorizontalBarChart from './charts/_HorizontalBarChart'
-import { getStateName, deathsBarColor } from './_utils'
+import HorizontalBarChart from '../../components/charts/horizontal-bar-chart'
+import { getStateName, deathsBarColor } from '../../utilities/visualization'
 
 import './state-cumulative-deaths-container.scss'
