+++ src/pages/dashboard/map-container.scss
@@ -82,10 +82,10 @@
 .legend-text {
   border-bottom: solid 3px;
   &.total {
-    border-color: $color-plum-300;
+    border-color: $color-plum-500;
   }
   &.positive {
-    border-color: $color-honey-500;
+    border-color: $color-honey-400;
   }
 }
 .add-margin-small {
